ui improvements

+++ scxrd/views/measurement_views.py
@@ -200,7 +200,7 @@ class MeasurementEditView(LoginRequiredMixin, UpdateView):
         request.POST._mutable = True
         request.POST['number'] = self.object.number
         request.POST._mutable = False
-        sample = self.object.sample
+        #sample = self.object.sample
         final = self.object.final
         form: MeasurementEditForm = self.get_form()
         if final:
@@ -210,17 +210,17 @@ class MeasurementEditView(LoginRequiredMixin, UpdateView):
         if form.is_valid():
             print('Form is valid')
             exp: Measurement = form.save(commit=False)
-            # Otherwise sample id gets lost: why?
-            exp.sample = sample
+            exp.refresh_from_db()
             if form.cleaned_data.get('final'):
                 if self.all_files_there(form):
                     exp.final = form.cleaned_data.get('final')
                 else:
-                    messages.warning(request,
-                                     _(
-                                         'You can only finalize a measurement after uploading a CIF, report and checkcif file!'))
+                    messages.warning(request, _('You can only finalize a measurement after uploading '
+                                                'a CIF, report and checkcif file!'))
                     return self.form_invalid(form)
-            exp.operator = request.user
+            if not exp.operator == request.user: # and not request.user.is_superuser:
+                messages.warning(request, _('You can only edit your own measurements.'))
+                return self.form_invalid(form)
             if request.POST.get('cif_file_on_disk-clear'):
                 exp.ciffilemodel.delete()
             if request.POST.get('checkcif_on_disk-clear'):
@@ -350,5 +350,5 @@ class MeasurementsListJsonUser(MeasurementListJson):
         qs = super().filter_queryset(qs)
         return qs.filter(Q(operator=User.objects.get(username=self.user)))
 
-    #def render_column(self, row, column):
+    # def render_column(self, row, column):
     #    return super()._render_column(row, column)
